Refactor boosters API to include catalog context.
Updated routes and API calls to manage boosters within the context of their catalog. Introduced `boostersIndex` tracking in the store and updated components to use the enhanced structure for filtering, editing, and deleting boosters.

+++ src/api/boosters.ts
@@ -27,12 +27,12 @@ export async function getBoosterByIdApi(boosterId: string): Promise<Response> {
   });
 }
 
-export async function createBoosterApi(booster: {
-  name: string;
-  description: string;
-  isActive: boolean;
-}): Promise<Response> {
-  return fetch(BASE_URL_BOOSTERS, {
+export async function createBoosterApi(booster: Booster, catalogId: string | undefined): Promise<Response> {
+  if (!catalogId) {
+    throw new Error("Index is required to create a booster.");
+  }
+
+  return fetch(`${BASE_URL_BOOSTERS}/${catalogId}/boosters`, {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
@@ -42,10 +42,7 @@ export async function createBoosterApi(booster: {
   });
 }
 
-export async function updateBoosterApi(
-  boosterId: string,
-  updatedData: Booster
-): Promise<Response> {
+export async function updateBoosterApi(booster: Booster, index: number): Promise<Response> {
   const url = `${BASE_URL_BOOSTERS}/${boosterId}`;
   return fetch(url, {
     method: "PUT",
@@ -53,13 +50,16 @@ export async function updateBoosterApi(
       "Content-Type": "application/json",
       "Authorization": `Bearer ${getAuthToken()}`
     },
-    body: JSON.stringify(updatedData)
+    body: JSON.stringify(booster)
   });
 }
 
 //TODO
-export async function deleteBoosterApi(boosterId: string): Promise<Response> {
-  const url = `${BASE_URL_BOOSTERS}/${boosterId}`;
+export async function deleteBoosterApi(boosterIndex: string | undefined,
+                                       catalogId: string | undefined): Promise<Response> {
+  if (!boosterIndex || !catalogId) throw new Error("boosterIndex and catalogId are required to delete a booster.");
+
+  const url = `${BASE_URL_BOOSTERS}/${catalogId}/boosters/${boosterIndex}`;
   return fetch(url, {
     method: "DELETE",
     headers: {
